trainer/auto_retrain.py
import joblib
import pandas as pd
from sqlalchemy.orm import Session
from database.models import SessionLocal, AegisLog, HistoricoTrade
from sklearn.ensemble import RandomForestClassifier

class AegisTrainer:
    """O Treinador: Aprende com os erros e acertos registrados no banco."""

    def __init__(self, model_path='brain/models/modelo_trading.pkl'):
        self.model_path = model_path

    def collect_training_data(self):
        session = SessionLocal()
        # Busca logs que resultaram em trade finalizado
        logs = session.query(AegisLog).all()
        trades = session.query(HistoricoTrade).all()
        
        if not logs or not trades:
            print("📉 [TRAINER] Dados insuficientes para re-treinamento.")
            return None, None

        # 1. Transforma logs técnicos em DataFrame de Features (X)
        # 2. Transforma o resultado do HistoricoTrade em Labels (y) -> 1 se lucro, 0 se prejuízo
        # Lógica de processamento de dados...
        
        session.close()
        return logs, trades

    def train_evolution(self):
        print("🔄 [TRAINER] Iniciando ciclo de auto-aperfeiçoamento...")
        X, y = self.collect_training_data()
        
        if X is not None:
            # Cria um novo modelo com os dados atualizados
            model = RandomForestClassifier(n_estimators=100)
            # model.fit(X, y)
            
            # Salva por cima do modelo antigo (Evolução)
            # joblib.dump(model, self.model_path)
            print(f"🏆 [TRAINER] Novo modelo Aegis Alpha forjado e salvo em {self.model_path}")