requirements.txt
alembic==1.18.4
fastapi==0.135.3
httpx==0.28.1
joblib==1.5.3
matplotlib==3.10.8
numpy==2.4.4
pandas==3.0.2
pyTelegramBotAPI==4.32.0
python-dotenv==1.2.2
python_binance==1.0.36
scikit_learn==1.8.0
SQLAlchemy==2.0.49
psycopg2-binary
xgboost
